picture_sim_python_app/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
julia==0.6.2
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
tqdm==4.67.1
watchdog==6.0.0
